beginning stages of adding I2C functionality

+++ Controller Modules/manualControl/manualControl/UART.c
@@ -18,7 +18,7 @@ void init_USART(void)
 }
 
 
-void USART_transmit(uint32_t data)
+void USART_transmit(uint16_t data)
 {
 	int nDigits = log10(data)+1;
 	int numParser;
